fix: restructure smooth trace withh fallback logic
