+++ src/model/smooth.ts
@@ -1,6 +1,5 @@
-import type { LinePoint, MaidrLayer, SmoothPoint } from '@type/grammar';
+import type { MaidrLayer } from '@type/grammar';
 import type { AudioState } from '@type/state';
-import { Svg } from '@util/svg';
 import { LineTrace } from './line';
 
 export class SmoothTrace extends LineTrace {
@@ -29,45 +28,5 @@ export class SmoothTrace extends LineTrace {
       // Only use groupIndex if there are multiple lines (actual multiline smooth plot)
       ...this.getAudioGroupIndex(),
     };
-  }
-
-  // Helper to access data points (since points is private in LineTrace)
-  protected getDataPoints(row: number): LinePoint[] {
-    return this.points?.[row] || [];
-  }
-
-  protected mapToSvgElements(selectors?: string[]): SVGElement[][] | null {
-    if (!selectors || selectors.length !== this.lineValues.length) {
-      return null;
-    }
-
-    const svgElements: SVGElement[][] = [];
-    let allFailed = true;
-    for (let r = 0; r < selectors.length; r++) {
-      const lineElement = Svg.selectElement(selectors[r], false);
-      if (!lineElement) {
-        svgElements.push([]);
-        continue;
-      }
-
-      // Use data points directly for highlight circles
-      const dataPoints = this.getDataPoints(r) as SmoothPoint[];
-      const linePointElements: SVGElement[] = [];
-      for (const pt of dataPoints) {
-        if (typeof pt.svg_x === 'number' && typeof pt.svg_y === 'number') {
-          linePointElements.push(Svg.createCircleElement(pt.svg_x, pt.svg_y, lineElement));
-        }
-      }
-
-      if (linePointElements.length > 0) {
-        allFailed = false;
-      }
-      svgElements.push(linePointElements);
-    }
-
-    if (allFailed) {
-      return null;
-    }
-    return svgElements;
   }
 }
