+++ src/model/smoothtraceFactory.ts
@@ -1,0 +1,20 @@
+import type { MaidrLayer } from '@type/grammar';
+import { SmoothTrace } from './smooth';
+import { SmoothTraceSvgXY } from './smoothSvgXY';
+
+function isSmoothPoint(pt: any): pt is { svg_x: number; svg_y: number } {
+  return typeof pt?.svg_x === 'number' && typeof pt?.svg_y === 'number';
+}
+
+export function createSmoothTrace(layer: MaidrLayer): SmoothTrace | SmoothTraceSvgXY {
+  // If the data has svg_x/svg_y, use the special class
+  const hasSvgXY = Array.isArray(layer.data)
+    && layer.data.length > 0
+    && Array.isArray(layer.data[0])
+    && isSmoothPoint(layer.data[0][0]);
+
+  if (hasSvgXY) {
+    return new SmoothTraceSvgXY(layer);
+  }
+  return new SmoothTrace(layer);
+}
